refactor(archtest): Persistence/test-infra/misc 规则族迁 module-path-agnostic CellRule + 缩 baseline [#1640]
M3 PR-9（EPIC #1302）。10 个「Persistence/test-infra/misc」族 archtest 规则迁
module-path-agnostic：规则逻辑（const/扫描 helper/Check*）从 _test.go MOVE 进非 test .go
（CellRule/Check*/Diagnostic 形态，可被外部 Cell 仓 import 运行），_test.go slim 为
dogfood（Report(t,id,Check*)）+ RED fixture + 盲区反向自检。平台符号路径全部派生
PlatformModulePath，不再硬编码 github.com/ghbvf/gocell；行为零变（AI-robust 评级 godoc 原样搬运）。
对标兄弟批 #1632/#1635/#1638/#1639。

规则（各新 .go + slim _test.go）：
- clock_invariants（KERNEL-CLOCK-LEAF-FALLBACK / -RESET-RELATIVE-PROD / PROD-CLOCK-INJECTION /
  CLOCK-POSITIONAL-INJECTION，4 子规则）
- security_defaults（SEC-FAIL-CLOSED-02..10）
- celltest_import_boundary（CELLTEST-IMPORT-BOUNDARY-01）
- celltest_import_scope（CELLTEST-IMPORT-SCOPE-01；表驱动用例平台路径亦派生 PlatformModulePath）
- required_dep_nil_guard（REQUIRED-DEP-NIL-GUARD-01）
- audit_hash_input_frozen（AUDIT-HASH-INPUT-FROZEN-01）
- distlock_lock_not_context（DISTLOCK-LOCK-NOT-CONTEXT-01）
- pg_repo_ambient_tx（PG-REPO-AMBIENT-TX-01）
- notfound_test_strict（POSTGRES-NOTFOUND-TEST-OTHER-ERROR-MIXUP-ARCHTEST-01）
- test_polling_external_reason_literal（TEST-POLLING-EXTERNAL-REASON-LITERAL-01）

baseline ratchet-down：本批 10 文件名从 module_path_funnel.baseline 删除（按文件名匹配）。
funnel 扫 _test.go + 新 .go，残留 bare 字面量即 CI 红——漏迁不可 merge，推进 #1304 Hard 终态。

register=none for all 10（allowlist/scan-scope gocell-hardcoded，无 ConfigForExternalCell
consumer-extension；外部仓 vacuous pass，各 .go godoc 写明理由）。

复杂度：移入 .go 的扫描函数受 gocognit/cyclop 约束（.golangci.yml 仅 exclude _test.go）；
线性 type-resolution 序列用 author-approved //nolint:gocognit[,cyclop] R2 标注，非结构性拆分。

验证：go build ./... / go test ./tools/archtest（10 规则 + funnel + platform-guard 全绿）/
golangci-lint run ./tools/archtest（0 issues）。
