+++ tools/archtest/celltest_import_scope_test.go
@@ -8,50 +8,19 @@
 package archtest
 
 import (
-	"fmt"
-	"path/filepath"
-	"regexp"
-	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
-// celltestImportPathPattern matches module-relative import paths whose final
-// segment is `[a-z]+test` under cells/{X}/ (e.g. cells/configcore/configcoretest).
-// The trailing `$` anchors the match to the top-level testutil package — nested
-// sub-packages (cells/configcore/configcoretest/sub) are intentionally NOT
-// matched here; they should be ban-covered by their parent package's
-// boundary, which an importer must traverse through the matched root first.
-var celltestImportPathPattern = regexp.MustCompile(`^github\.com/[^/]+/[^/]+/cells/[a-z]+/[a-z]+test$`)
-
-// isCellTestImportPath reports whether an absolute import path matches the
-// cells/{X}/{X}test naming convention.
-//
-// Blind spots: the pattern matches the full path literal from the import
-// declaration; it does not follow type aliases or build-tag-gated imports.
-// Both of those forms are effectively impossible to use as a production import
-// without also triggering a Go compilation error, so they are acceptable
-// non-coverage.
-func isCellTestImportPath(importPath string) bool {
-	return celltestImportPathPattern.MatchString(importPath)
-}
-
-// TestCelltestImportScope enforces CELLTEST-IMPORT-SCOPE-01:
-//
-// No production Go file (i.e. NOT *_test.go and NOT in a test-infrastructure
-// directory) may import a package whose import path matches
-// cells/[a-z]+/[a-z]+test$. Those packages are cell-level testutil packages
-// and importing them from production code would embed test fixtures and
-// t-bound helpers in a release binary, signaling a layering mistake.
+// TestCelltestImportScope enforces CELLTEST-IMPORT-SCOPE-01 via
+// CheckCelltestImportScope (celltest_import_scope.go). The scanner logic
+// lives in the non-test file so external Cell repositories can import and run
+// the rule without needing GoCell's _test.go compilation.
 //
 // Complementary rules:
 //   - TESTUTIL-BOUNDARY-01 (testutil_boundary_test.go): guards "testutil" segment paths.
 //   - This rule: guards the cells/*/*test/ naming form (e.g. cells/configcore/configcoretest).
-//
-// The check is discovery-based: any new cells/{X}/{X}test/ package is
-// automatically covered without further edits to this file.
 //
 // Blind spots (non-coverage outside declared rule scope):
 //   - Import aliases that rename the package after import: these require type
@@ -60,51 +29,7 @@ func isCellTestImportPath(importPath string) bool {
 //   - Build-tag-gated imports (//go:build ignore) that never compile: these
 //     are not production imports in practice and do not reach a release binary.
 func TestCelltestImportScope(t *testing.T) {
-	root := findModuleRoot(t)
-	modPath := readModulePath(t, root)
-
-	allGoFiles, err := collectGoFiles(root)
-	require.NoError(t, err, "failed to collect .go files")
-	require.NotEmpty(t, allGoFiles, "no .go files found — module root may be wrong")
-
-	var violations []string
-	for _, f := range allGoFiles {
-		// Skip _test.go — they are permitted to import test-infra packages.
-		if strings.HasSuffix(f, "_test.go") {
-			continue
-		}
-		rel, err := filepath.Rel(root, f)
-		require.NoError(t, err)
-		rel = filepath.ToSlash(rel)
-		// Skip test-infrastructure paths (outboxtest/, configcoretest/, etc.)
-		// — they may import sibling test packages.
-		if isTestInfraPath(rel) {
-			continue
-		}
-
-		imports, err := parseImports(f)
-		require.NoError(t, err, "failed to parse %s", f)
-		for _, imp := range imports {
-			if !strings.HasPrefix(imp, modPath+"/") {
-				continue
-			}
-			if isCellTestImportPath(imp) {
-				violations = append(violations,
-					fmt.Sprintf("CELLTEST-IMPORT-SCOPE-01: %s (production file) imports %s "+
-						"(cells/*/*test packages are test-infrastructure; only *_test.go or "+
-						"test-infra packages may import them)", rel, imp))
-			}
-		}
-	}
-
-	if len(violations) > 0 {
-		for _, v := range violations {
-			t.Logf("%s", v)
-		}
-	}
-	assert.Empty(t, violations,
-		"production (non-_test.go, non-test-infra) files must not import cells/*/*test packages; "+
-			"these packages embed test fixtures — import them only from *_test.go or test-infrastructure files")
+	Report(t, "CELLTEST-IMPORT-SCOPE-01", CheckCelltestImportScope(t, ConfigForExternalCell{}))
 }
 
 // TestCelltestImportPath_PatternTable is a table-driven unit test for the
@@ -122,53 +47,53 @@ func TestCelltestImportPath_PatternTable(t *testing.T) {
 		// Positive: should match (violation)
 		{
 			name:      "configcoretest direct",
-			path:      "github.com/ghbvf/gocell/cells/configcore/configcoretest",
+			path:      PlatformModulePath + "/cells/configcore/configcoretest",
 			wantMatch: true,
 		},
 		{
 			name:      "accesscoretest direct",
-			path:      "github.com/ghbvf/gocell/cells/accesscore/accesscoretest",
+			path:      PlatformModulePath + "/cells/accesscore/accesscoretest",
 			wantMatch: true,
 		},
 		{
 			name:      "auditcoretest direct",
-			path:      "github.com/ghbvf/gocell/cells/auditcore/auditcoretest",
+			path:      PlatformModulePath + "/cells/auditcore/auditcoretest",
 			wantMatch: true,
 		},
 		// Negative: should NOT match (compliant / out-of-scope)
 		{
 			name:      "production cell package",
-			path:      "github.com/ghbvf/gocell/cells/configcore/slices/flagread",
+			path:      PlatformModulePath + "/cells/configcore/slices/flagread",
 			wantMatch: false,
 		},
 		{
 			name:      "kernel package",
-			path:      "github.com/ghbvf/gocell/kernel/outbox",
+			path:      PlatformModulePath + "/kernel/outbox",
 			wantMatch: false,
 		},
 		{
 			name:      "outboxtest (kernel, not cells)",
-			path:      "github.com/ghbvf/gocell/kernel/outbox/outboxtest",
+			path:      PlatformModulePath + "/kernel/outbox/outboxtest",
 			wantMatch: false,
 		},
 		{
 			name:      "testutil path (covered by TESTUTIL-BOUNDARY-01)",
-			path:      "github.com/ghbvf/gocell/cells/configcore/internal/testutil",
+			path:      PlatformModulePath + "/cells/configcore/internal/testutil",
 			wantMatch: false,
 		},
 		{
 			name:      "sub-package of celltest (blind spot — outside declared scope)",
-			path:      "github.com/ghbvf/gocell/cells/configcore/configcoretest/sub",
+			path:      PlatformModulePath + "/cells/configcore/configcoretest/sub",
 			wantMatch: false,
 		},
 		{
 			name:      "runtime package",
-			path:      "github.com/ghbvf/gocell/runtime/auth",
+			path:      PlatformModulePath + "/runtime/auth",
 			wantMatch: false,
 		},
 		{
 			name:      "examples cell package",
-			path:      "github.com/ghbvf/gocell/examples/todoorder/cells/ordercore",
+			path:      PlatformModulePath + "/examples/todoorder/cells/ordercore",
 			wantMatch: false,
 		},
 		{
